+++ .eslintrc.cjs
@@ -1,3 +1,14 @@
+const TAILWIND_CONFIG = {
+  extends: ['plugin:tailwindcss/recommended'],
+  rules: {
+    'tailwindcss/classnames-order': 'error',
+    'tailwindcss/enforces-negative-arbitrary-values': 'error',
+    'tailwindcss/enforces-shorthand': 'error',
+    'tailwindcss/migration-from-tailwind-2': 'error',
+    'tailwindcss/no-custom-classname': 'error'
+  }
+}
+
 module.exports = {
   root: true,
   parser: '@typescript-eslint/parser',
@@ -35,6 +46,56 @@ module.exports = {
             ]
           }
         ]
+      }
+    },
+    {
+      ...TAILWIND_CONFIG,
+      files: 'packages/nextra-theme-docs/**/*',
+      settings: {
+        tailwindcss: {
+          config: 'packages/nextra-theme-docs/tailwind.config.js',
+          callees: ['cn'],
+          whitelist: ['nextra-breadcrumb', 'nextra-callout', 'nextra-bleed']
+        }
+      }
+    },
+    {
+      ...TAILWIND_CONFIG,
+      files: 'packages/nextra-theme-blog/**/*',
+      settings: {
+        tailwindcss: {
+          config: 'packages/nextra-theme-blog/tailwind.config.js',
+          whitelist: ['subheading-', 'post-item', 'post-item-more']
+        }
+      }
+    },
+    {
+      ...TAILWIND_CONFIG,
+      files: 'packages/nextra/**/*',
+      settings: {
+        tailwindcss: {
+          config: 'packages/nextra-theme-docs/tailwind.config.js'
+        }
+      }
+    },
+    {
+      ...TAILWIND_CONFIG,
+      files: 'examples/swr-site/**/*',
+      settings: {
+        tailwindcss: {
+          config: 'examples/swr-site/tailwind.config.js'
+        }
+      }
+    },
+    {
+      ...TAILWIND_CONFIG,
+      files: 'docs/**/*',
+      settings: {
+        tailwindcss: {
+          config: 'docs/tailwind.config.js',
+          callees: ['cn'],
+          whitelist: ['dash-ring', 'theme-1', 'theme-2', 'theme-3', 'theme-4', 'border-primary-100/10$']
+        }
       }
     }
   ]