+++ packages/nextra-theme-docs/src/components/callout.tsx
@@ -11,14 +11,19 @@ const TypeToEmoji = {
 
 type CalloutType = keyof typeof TypeToEmoji
 
-const themes: Record<CalloutType, string> = {
-  default:
-    'nx-bg-orange-50 nx-border-orange-100 nx-text-orange-800 dark:nx-text-orange-300 dark:nx-bg-orange-400/20 dark:nx-border-orange-400/30',
-  error:
-    'nx-bg-red-100 nx-border-red-200 nx-text-red-900 dark:nx-text-red-200 dark:nx-bg-red-900/30 dark:nx-border-red-200/30',
-  info: 'nx-bg-blue-100 nx-border-blue-200 nx-text-blue-900 dark:nx-text-blue-200 dark:nx-bg-blue-900/30 dark:nx-border-blue-200/30',
-  warning:
-    'nx-bg-yellow-50 nx-border-yellow-100 nx-text-yellow-900 dark:nx-text-yellow-200 dark:nx-bg-yellow-700/30'
+const classes: Record<CalloutType, string> = {
+  default: cn(
+    'nx-border-orange-100 nx-bg-orange-50 nx-text-orange-800 dark:nx-border-orange-400/30 dark:nx-bg-orange-400/20 dark:nx-text-orange-300'
+  ),
+  error: cn(
+    'nx-border-red-200 nx-bg-red-100 nx-text-red-900 dark:nx-border-red-200/30 dark:nx-bg-red-900/30 dark:nx-text-red-200'
+  ),
+  info: cn(
+    'nx-border-blue-200 nx-bg-blue-100 nx-text-blue-900 dark:nx-border-blue-200/30 dark:nx-bg-blue-900/30 dark:nx-text-blue-200'
+  ),
+  warning: cn(
+    'nx-border-yellow-100 nx-bg-yellow-50 nx-text-yellow-900 dark:nx-bg-yellow-700/30 dark:nx-text-yellow-200'
+  )
 }
 
 type CalloutProps = {
@@ -35,9 +40,9 @@ export function Callout({
   return (
     <div
       className={cn(
-        'nextra-callout nx-border nx-mt-6 nx-flex nx-rounded-lg nx-py-2 ltr:nx-pr-4 rtl:nx-pl-4',
+        'nextra-callout nx-mt-6 nx-flex nx-rounded-lg nx-border nx-py-2 ltr:nx-pr-4 rtl:nx-pl-4',
         'contrast-more:nx-border-current contrast-more:dark:nx-border-current',
-        themes[type]
+        classes[type]
       )}
     >
       <div