+++ packages/nextra-theme-docs/src/components/toc.tsx
@@ -49,7 +49,7 @@ export function TOC({ headings, filePath }: TOCProps): ReactElement {
 
   useEffect(() => {
     if (!activeSlug) return
-    const anchor = tocRef.current?.querySelector(`li > a[href="#${activeSlug}"`)
+    const anchor = tocRef.current?.querySelector(`li > a[href="#${activeSlug}"]`)
 
     if (anchor) {
       scrollIntoView(anchor, {
@@ -71,8 +71,8 @@ export function TOC({ headings, filePath }: TOCProps): ReactElement {
     <div
       ref={tocRef}
       className={cn(
-        'nextra-scrollbar nx-sticky nx-top-16 nx-overflow-y-auto nx-pr-4 nx-pt-8 nx-text-sm nx-[hyphens:auto]',
-        'ltr:-nx-mr-4 rtl:-nx-ml-4 nx-max-h-[calc(100vh-var(--nextra-navbar-height)-env(safe-area-inset-bottom))]'
+        'nextra-scrollbar nx-sticky nx-top-16 nx-overflow-y-auto nx-pr-4 nx-pt-8 nx-text-sm [hyphens:auto]',
+        'nx-max-h-[calc(100vh-var(--nextra-navbar-height)-env(safe-area-inset-bottom))] ltr:-nx-mr-4 rtl:-nx-ml-4'
       )}
     >
       {hasHeadings && (
@@ -113,8 +113,8 @@ export function TOC({ headings, filePath }: TOCProps): ReactElement {
           className={cn(
             hasHeadings &&
               'nx-mt-8 nx-border-t nx-bg-white nx-pt-8 nx-shadow-[0_-12px_16px_white] dark:nx-bg-dark dark:nx-shadow-[0_-12px_16px_#111]',
-            'nx-sticky nx-bottom-0 nx-pb-8 dark:nx-border-neutral-800 nx-flex nx-flex-col nx-items-start nx-gap-2',
-            'contrast-more:nx-shadow-none contrast-more:nx-border-t contrast-more:nx-border-neutral-400 contrast-more:dark:nx-border-neutral-400'
+            'nx-sticky nx-bottom-0 nx-flex nx-flex-col nx-items-start nx-gap-2 nx-pb-8 dark:nx-border-neutral-800',
+            'contrast-more:nx-border-t contrast-more:nx-border-neutral-400 contrast-more:nx-shadow-none contrast-more:dark:nx-border-neutral-400'
           )}
         >
           {config.feedback.content ? (