+++ packages/nextra-theme-docs/src/constants.tsx
@@ -76,8 +76,8 @@ export const DEFAULT_THEME: DocsThemeConfig = {
   i18n: [],
   logo: (
     <>
-      <span className="font-extrabold">Nextra</span>
-      <span className="ml-2 hidden font-normal text-gray-600 md:inline">
+      <span className="nx-font-extrabold">Nextra</span>
+      <span className="nx-ml-2 nx-hidden nx-font-normal nx-text-gray-600 md:nx-inline">
         The Next Docs Builder
       </span>
     </>
@@ -121,15 +121,22 @@ export const DEFAULT_THEME: DocsThemeConfig = {
       )
     },
     emptyResult: (
-      <span className="block select-none p-8 text-center text-sm text-gray-400">
+      <span className="nx-block nx-select-none nx-p-8 nx-text-center nx-text-sm nx-text-gray-400">
         No results found.
       </span>
     ),
+    loading() {
+      const { locale } = useRouter()
+      if (locale === 'zh-CN') return '正在加载…'
+      if (locale === 'ru') return 'Загрузка…'
+      if (locale === 'fr') return 'Сhargement…'
+      return 'Loading…'
+    },
     placeholder() {
       const { locale } = useRouter()
       if (locale === 'zh-CN') return '搜索文档…'
-      if (locale === 'ru-RU') return 'Поиск документации…'
-      if (locale === 'fr-FR') return 'Rechercher de la documentation…'
+      if (locale === 'ru') return 'Поиск документации…'
+      if (locale === 'fr') return 'Rechercher de la documentation…'
       return 'Search documentation…'
     }
   },
@@ -138,7 +145,7 @@ export const DEFAULT_THEME: DocsThemeConfig = {
     labels: 'bug'
   },
   sidebar: {
-    defaultMenuCollapsed: false,
+    defaultMenuCollapseLevel: 2,
     titleComponent: ({ title }) => <>{title}</>
   },
   toc: {
@@ -165,7 +172,7 @@ export const LEGACY_CONFIG_OPTIONS: Record<string, string> = {
   bannerKey: 'banner.key',
   bodyExtraContent: 'main',
   customSearch: 'search.component',
-  defaultMenuCollapsed: 'sidebar.defaultMenuCollapsed',
+  defaultMenuCollapsed: 'sidebar.defaultMenuCollapseLevel',
   feedbackLabels: 'feedback.labels',
   feedbackLink: 'feedback.content',
   floatTOC: 'toc.float',