+++ packages/nextra-theme-docs/src/mdx-components.tsx
@@ -110,7 +110,7 @@ const createHeaderLink = (
         className={cn(
           'nx-font-semibold nx-tracking-tight',
           {
-            h2: 'nx-mt-10 nx-text-3xl nx-border-b nx-pb-1 dark:nx-border-primary-100/10 contrast-more:nx-border-neutral-400 contrast-more:dark:nx-border-neutral-400',
+            h2: 'nx-mt-10 nx-border-b nx-pb-1 nx-text-3xl contrast-more:nx-border-neutral-400 dark:nx-border-primary-100/10 contrast-more:dark:nx-border-neutral-400',
             h3: 'nx-mt-8 nx-text-2xl',
             h4: 'nx-mt-8 nx-text-xl',
             h5: 'nx-mt-8 nx-text-lg',
@@ -180,7 +180,7 @@ const Details = ({
 
   return (
     <details
-      className="nx-my-4 nx-rounded nx-border nx-border-gray-200 nx-bg-white nx-p-2 nx-shadow-sm dark:nx-border-neutral-800 dark:nx-bg-neutral-900 first:nx-mt-0 last:nx-mb-0"
+      className="nx-my-4 nx-rounded nx-border nx-border-gray-200 nx-bg-white nx-p-2 nx-shadow-sm first:nx-mt-0 dark:nx-border-neutral-800 dark:nx-bg-neutral-900"
       {...props}
       open={delayedOpenState}
       {...(openState && { 'data-expanded': true })}
@@ -196,9 +196,9 @@ const Summary = (props: ComponentProps<'summary'>): ReactElement => {
   return (
     <summary
       className={cn(
-        'nx-list-none nx-cursor-pointer nx-p-1 nx-transition-colors hover:nx-bg-gray-100 dark:hover:nx-bg-neutral-800',
-        "before:nx-mr-1 before:content-[''] before:nx-inline-block before:nx-transition-transform dark:before:nx-invert",
-        '[[data-expanded]>&]:before:nx-rotate-90 rtl:before:nx-rotate-180'
+        'nx-cursor-pointer nx-list-none nx-p-1 nx-transition-colors hover:nx-bg-gray-100 dark:hover:nx-bg-neutral-800',
+        "before:nx-mr-1 before:nx-inline-block before:nx-transition-transform before:nx-content-[''] dark:before:nx-invert",
+        'rtl:before:nx-rotate-180 [[data-expanded]>&]:before:nx-rotate-90'
       )}
       {...props}
       onClick={e => {
@@ -239,13 +239,13 @@ export const getComponents = ({
     h6: createHeaderLink('h6', context),
     ul: (props: ComponentProps<'ul'>) => (
       <ul
-        className="ltr:nx-ml-6 rtl:nx-mr-6 nx-mt-6 nx-list-disc first:nx-mt-0"
+        className="nx-mt-6 nx-list-disc first:nx-mt-0 ltr:nx-ml-6 rtl:nx-mr-6"
         {...props}
       />
     ),
     ol: (props: ComponentProps<'ol'>) => (
       <ol
-        className="ltr:nx-ml-6 rtl:nx-mr-6 nx-mt-6 nx-list-decimal"
+        className="nx-mt-6 nx-list-decimal first:nx-mt-0 ltr:nx-ml-6 rtl:nx-mr-6"
         {...props}
       />
     ),
@@ -253,8 +253,8 @@ export const getComponents = ({
     blockquote: (props: ComponentProps<'blockquote'>) => (
       <blockquote
         className={cn(
-          'nx-mt-6 first:nx-mt-0 nx-border-gray-300 nx-italic nx-text-gray-700 dark:nx-border-gray-700 dark:nx-text-gray-400',
-          'rtl:nx-border-r-2 rtl:nx-pr-6 ltr:nx-border-l-2 ltr:nx-pl-6'
+          'nx-mt-6 nx-border-gray-300 nx-italic nx-text-gray-700 dark:nx-border-gray-700 dark:nx-text-gray-400',
+          'first:nx-mt-0 ltr:nx-border-l-2 ltr:nx-pl-6 rtl:nx-border-r-2 rtl:nx-pr-6'
         )}
         {...props}
       />
@@ -270,12 +270,12 @@ export const getComponents = ({
     ),
     table: (props: ComponentProps<'table'>) => (
       <Table
-        className="nextra-scrollbar nx-mt-6 first:nx-mt-0 nx-p-0"
+        className="nextra-scrollbar nx-mt-6 nx-p-0 first:nx-mt-0"
         {...props}
       />
     ),
     p: (props: ComponentProps<'p'>) => (
-      <p className="nx-mt-6 first:nx-mt-0 nx-leading-7" {...props} />
+      <p className="nx-mt-6 nx-leading-7 first:nx-mt-0" {...props} />
     ),
     tr: Tr,
     th: Th,