+++ packages/nextra/src/components/th.tsx
@@ -2,7 +2,7 @@ import React, { ComponentProps } from 'react'
 
 export const Th = (props: ComponentProps<'th'>) => (
   <th
-    className="m-0 border border-gray-300 px-4 py-2 dark:border-gray-600 font-semibold"
+    className="nx-m-0 nx-border nx-border-gray-300 nx-px-4 nx-py-2 nx-font-semibold dark:nx-border-gray-600"
     {...props}
   />
 )